+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+1.0.0 (unreleased)
+-----
+* [FEATURE] Added new node <system_configuration> for specifying configuration settings in a more readable manner.
+* [FEATURE] Added backend overview which shows the configured environments and which commands will be executed where.
+* [FEATURE] Added new command stages 'pre_configure' and 'post_configure'.
+* [FEATURE] Added n98-magerun command db:truncate.
+* [FEATURE] Added n98-magerun command ls:env:reset. 
+
 0.0.1 (09.04.2014)
 -----
 * Initial release